Backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
uvicorn[standard]